Use mask/unmask for irq and _ack rather than EOI

+++ kern/arch/x86_64/lapic.c
@@ -61,7 +61,7 @@ init_lapic ()
 }
 
 void
-send_eoi (uint8_t irq)
+send_eoi (int irq)
 {
   (void)irq;
